requirements.txt
alembic==1.13.1
aniso8601==9.0.1
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
Flask==3.0.3
Flask-Cors==4.0.1
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
flask-query-builder==0.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
mysqlclient==2.2.4
packaging==24.1
passlib==1.7.4
pycparser==2.22
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
Werkzeug==3.0.3
